[DWP] Stop storing DWPStringPool strings in Buffer
When we can just write it out. This decreases peak memory usage of llvm-dwp on clang debug info from ~800MB to ~350MB without measurable performance changes.

Also remove DWPStringPool.h to avoid circular headers. It wasn't doing anything anyway...
