+++ llvm/include/llvm/DWP/DWP.h
@@ -1,7 +1,6 @@
 #ifndef LLVM_DWP_DWP_H
 #define LLVM_DWP_DWP_H
 
-#include "DWPStringPool.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/MapVector.h"
 #include "llvm/ADT/SmallVector.h"
@@ -139,6 +138,28 @@ public:
   Error writeWASM(raw_pwrite_stream &OS);
   Error write(raw_pwrite_stream &OS) {
     return IsWASM ? writeWASM(OS) : writeELF(OS);
+  }
+};
+
+class DWPStringPool {
+  DWPWriter &Out;
+  DenseMap<StringRef, uint64_t> Pool;
+  uint64_t Offset = 0;
+
+public:
+  DWPStringPool(DWPWriter &Out) : Out(Out) {}
+
+  uint64_t getOffset(const char *Str, unsigned Length) {
+    assert(strlen(Str) + 1 == Length && "Ensure length hint is correct");
+
+    StringRef Key(Str, Length);
+    auto Pair = Pool.insert(std::make_pair(Key, Offset));
+    if (Pair.second) {
+      Out.emitBytes(Key);
+      Offset += Length;
+    }
+
+    return Pair.first->second;
   }
 };
